seperate core machine parameter functions from parameter ui functions
