+++ TFT/src/User/Menu/MachineSettings.h
@@ -5,7 +5,4 @@
 void menuRGBSettings(void);
 void menuMachineSettings(void);
 void menuCustom(void);
-void saveEepromSettings(void);
-void restoreEepromSettings(void);
-void resetEepromSettings(void);
 #endif
